Client, Editor, Server, and Migrator all actually working after merging in hotfixes.
Former-commit-id: 3a443f7b57f5ee9a70dd46abbf4876f6ae8b8b79

+++ Intersect Migration Tool/UpgradeInstructions/Upgrade_11/Intersect_Convert_Lib/GameObjects/ClassBase.cs
@@ -133,8 +133,8 @@ namespace Intersect.Migration.UpgradeInstructions.Upgrade_11.Intersect_Convert_L
             PointIncrease = myBuffer.ReadInteger();
 
             //Exp Info
-            BaseExp = myBuffer.ReadLong();
-            ExpIncrease = myBuffer.ReadLong();
+            BaseExp = myBuffer.ReadInteger();
+            ExpIncrease = myBuffer.ReadInteger();
 
             //Regen
             for (int i = 0; i < (int) Vitals.VitalCount; i++)
